Move handlers in common/main instead of specific folder
